ml_pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
textblob>=0.17.1
nltk>=3.8.1
pyyaml>=6.0
joblib>=1.2.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
